Backend integration, UX

+++ backend/app/Http/Controllers/ChannelsController.php
@@ -13,7 +13,13 @@ class ChannelsController extends Controller
      */
     public function index()
     {
-        return response()->json(Channels::all());
+        $channels = Channels::all();
+
+        if (!$channels) {
+            return response()->json(['error' => 'Channels not found'], 404);
+        }
+
+        return response()->json($channels);
     }
 
     /**
@@ -21,19 +27,21 @@ class ChannelsController extends Controller
      */
     public function store(Request $request)
     {
-        $validated = $request->validate([
+        $request->validate([
             'name' => 'required',
-            'users_count' => 'required',
+            'clientsCount' => 'required',
         ]);
 
-        // take the data from body and create a new channel
+        if (Channels::where('name', $request->name)->exists()) {
+            return response()->json(['error' => 'Channel already exists'], 400);
+        }
 
         $newChannel = new Channels();
         $newChannel->name = $request->name;
-        $newChannel->users_count = $request->users_count;
+        $newChannel->clientsCount = $request->clientsCount;
         $newChannel->save();
         
-        return response()->json($newChannel);
+        return response()->json(['message' => 'Channel created successfully']);
 
     }
 
@@ -43,21 +51,31 @@ class ChannelsController extends Controller
     public function update(Request $request, Channels $channel)
     {
 
-        // new name or users_count required
-        if(!$request->name && !$request->users_count) {
-            return response()->json(['error' => 'name or users_count required'], 400);
-        }
+       $request->validate([
+           'name' => 'required',
+           'clientsCount' => 'required',
+       ]);
 
-        // check if the channel exists
+        // check if the updating channel exists
         if (!$channel) {
             return response()->json(['error' => 'Channel not found'], 404);
         }
 
+        // check if the different channel with the same name already exists
+        if ( Channels::where('name', $request->name)->where('id', '!=', $channel->id)->exists()) {
+            return response()->json(['error' => 'Channel with the same name already exists'], 400);
+        }
+
+        // check if the clientsCount is greater than 0
+        if ($request->clientsCount <= 0) {
+            return response()->json(['error' => 'clientsCount must be greater than 0'], 400);
+        }
+
         $channel->name = $request->name;
-        $channel->users_count = $request->users_count;
+        $channel->clientsCount = $request->clientsCount;
         $channel->save();
         
-        return response()->json($channel);
+        return response()->json(['message' => 'Channel updated successfully']);
     }
 
     /**
@@ -65,7 +83,12 @@ class ChannelsController extends Controller
      */
     public function destroy(Channels $channel)
     {
+
+        if (!$channel) {
+            return response()->json(['error' => 'Channel not found'], 404);
+        }
+
         $channel->delete();
-        return response()->json($channel);
+        return response()->json(['message' => 'Channel deleted successfully']);
     }
 }
